+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### New features
 
+* Added `Tag` methods `remove_class` and `add_style`. (#57)
+* Added support for `Tag`'s `add_class(prepend=)`. (#57)
+
 ### Bug fixes
 
 ### Other changes